Deliverable 5 : Refactor and latest version retrieval
