+++ storeroom/src/main/java/uk/ac/ebi/biosamples/jsonschema/jsonschemastore/schema/document/SchemaBlock.java
@@ -19,7 +19,7 @@ public class SchemaBlock {
   @Id private String id;
   private String schemaName;
   private String version;
-  private boolean isLatest;
+  private boolean latest;
   private String schema;
   @TextIndexed(weight = 2F)
   private String title;
